Changed Saving of Settings
Saving of settings was throwing errors in Mach4 and the debugger intermittently. It seemes as if the registeredSettings table in RapidChangeSettings.lua was not persisting.

Anyway, i changed it so the Wizard stores settings itself then calls on the module to save each setting when called to

TODO: USER INPUT VALIDATION